+++ Tests/Units/App/Libraries/Calendrier/Collection/Heure/Additionnelle.php
@@ -5,52 +5,16 @@ use App\Libraries\Calendrier\Collection\Heure\Additionnelle as _Additionnelle;
 
 /**
  * Classe de test des collections d'heures additionnelles
+ *
+ * @since  1.9
+ * @author Prytoegrian <prytoegrian@protonmail.com>
+ * @author Wouldsmina
+ *
  */
-class Additionnelle extends \Tests\Units\TestUnit
+class Additionnelle extends \Tests\Units\App\Libraries\Calendrier\Collection\AHeure
 {
-    public function beforeTestMethod($method)
+    protected function getTestedClass()
     {
-        parent::beforeTestMethod($method);
-        $this->result = new \mock\MYSQLIResult();
-        $this->db = new \mock\includes\SQL();
-        $this->calling($this->db)->query = $this->result;
-    }
-
-    private $db;
-    private $result;
-
-    public function testGetListeVoid()
-    {
-        $this->calling($this->result)->fetch_all = [];
-        $date = new \DateTimeImmutable();
-
-        $heures = new _Additionnelle($this->db, []);
-
-        $this->array($heures->getListe($date, $date, [], false))->isEmpty();
-    }
-
-    public function testGetListeFilled()
-    {
-        $this->calling($this->result)->fetch_all[1] = [['id' => 3]];
-        $this->calling($this->result)->fetch_all[2] = [[
-            'login' => 'Provencal le Gaulois',
-            'debut' => 1191929182,
-            'fin' =>   1199128919,
-            'statut' => \App\Models\AHeure::STATUT_VALIDATION_FINALE,
-        ],];
-
-        $heures = new _Additionnelle($this->db);
-
-        $nomComplet = 'Provencal le Gaulois';
-        $expected = [
-            '2007-10-09' => [[
-                'employe' => $nomComplet,
-                'statut' => \App\Models\AHeure::STATUT_VALIDATION_FINALE,
-            ]],
-        ];
-        $date = new \DateTimeImmutable();
-        $liste = $heures->getListe($date, $date, [], false);
-
-        $this->array($liste)->isIdenticalTo($expected);
+        return _Additionnelle::class;
     }
 }
